+++ tickplate.d.ts
@@ -1,3 +1,10 @@
-declare function tickplate(strings: Array<string>, ...keys: Array<string>): (values: object) => string;
+interface TickplateOptions {
+  delimiter?: any;
+}
+
+declare function tickplate(
+  strings: Array<string>,
+  ...keys: Array<string>
+): (values: object, opts?: TickplateOptions) => string;
 
 export = tickplate;